+++ CHANGELOG.md
@@ -38,7 +38,8 @@ All user visible changes to this project will be documented in this file. This p
     - Send `Event::RpcSettingsUpdated` when `Member` connects ([#75]);
     - Send relay mode in `Event::PeerCreated` which is used for configuring client's `RtcIceTransportPolicy` ([#79]);
     - Emit `TracksApplied` event to create new and update existing tracks ([#105]);
-    - `PeerConnection` renegotiation functionality ([#105]).
+    - `PeerConnection` renegotiation functionality ([#105]);
+    - Calculate and send call quality score based on RTC stats ([#132]).
 - [Coturn] integration:
     - [Coturn] sessions destroying ([#84]);
     - [Coturn] stats processing ([#94]).
@@ -73,6 +74,7 @@ All user visible changes to this project will be documented in this file. This p
 [#95]: /../../pull/95
 [#98]: /../../pull/98
 [#105]: /../../pull/105
+[#132]: /../../pull/132